Remove getDummyMessenger() Method
