+++ test/games/strategy/engine/chat/StatusTest.java
@@ -2,19 +2,30 @@ package games.strategy.engine.chat;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+import java.net.InetAddress;
 
 import org.junit.Test;
 
+import games.strategy.net.INode;
 import games.strategy.net.IServerMessenger;
 import games.strategy.net.Messengers;
-import games.strategy.triplea.delegate.MockObjects;
+import games.strategy.net.Node;
 import games.strategy.util.ThreadUtil;
 
 public class StatusTest {
 
   @Test
   public void testStatus() throws Exception {
-    final IServerMessenger messenger = MockObjects.getDummyMessenger();
+    IServerMessenger messenger = mock(IServerMessenger.class);
+    INode dummyNode = new Node("dummy", InetAddress.getLocalHost(), 0);
+    when(messenger.getLocalNode()).thenReturn(dummyNode);
+    when(messenger.getServerNode()).thenReturn(dummyNode);
+    when(messenger.isConnected()).thenReturn(true);
+    when(messenger.isServer()).thenReturn(true);
+
     final Messengers messengers = new Messengers(messenger);
     final StatusManager manager = new StatusManager(messengers);
     assertNull(manager.getStatus(messenger.getLocalNode()));
